chore: Reorganise API code and tests
