+++ tests/unit/core/api/test_index.js
@@ -1,20 +1,14 @@
 /* global window */
-import querystring from 'querystring';
-
 import config from 'config';
 import utf8 from 'utf8';
 
-import * as api from 'core/api';
-import { ADDON_TYPE_THEME, CLIENT_APP_ANDROID } from 'core/constants';
+import { callApi, createApiError, makeQueryString } from 'core/api';
 import {
   createApiResponse,
   createStubErrorHandler,
   generateHeaders,
-  signedInApiState,
   unexpectedSuccess,
-  userAuthToken,
 } from 'tests/unit/helpers';
-import { createFakeAutocompleteResult } from 'tests/unit/amo/helpers';
 
 
 describe(__filename, () => {
@@ -37,7 +31,7 @@ describe(__filename, () => {
         })
         .once()
         .returns(createApiResponse());
-      return api.callApi({ endpoint: 'resource', method: 'get' })
+      return callApi({ endpoint: 'resource', method: 'get' })
         .then(() => mockWindow.verify());
     });
 
@@ -49,7 +43,7 @@ describe(__filename, () => {
         })
         .once()
         .returns(createApiResponse());
-      return api.callApi({ endpoint }).then(() => mockWindow.verify());
+      return callApi({ endpoint }).then(() => mockWindow.verify());
     });
 
     it('clears an error handler before making a request', () => {
@@ -58,7 +52,7 @@ describe(__filename, () => {
       const errorHandler = createStubErrorHandler();
       sinon.stub(errorHandler, 'clear');
 
-      return api.callApi({ endpoint: 'resource', errorHandler })
+      return callApi({ endpoint: 'resource', errorHandler })
         .then(() => {
           expect(errorHandler.clear.called).toBeTruthy();
         });
@@ -74,7 +68,7 @@ describe(__filename, () => {
       const errorHandler = createStubErrorHandler();
       sinon.stub(errorHandler, 'handle');
 
-      return api.callApi({ endpoint: 'resource', errorHandler })
+      return callApi({ endpoint: 'resource', errorHandler })
         .then(unexpectedSuccess, () => {
           expect(errorHandler.handle.called).toBeTruthy();
           const args = errorHandler.handle.firstCall.args;
@@ -93,7 +87,7 @@ describe(__filename, () => {
       const errorHandler = createStubErrorHandler();
       sinon.stub(errorHandler, 'handle');
 
-      return api.callApi({ endpoint: 'resource' })
+      return callApi({ endpoint: 'resource' })
         .then(unexpectedSuccess, (err) => {
           expect(err.message).toEqual('pretend this was a response with invalid JSON');
         });
@@ -107,7 +101,7 @@ describe(__filename, () => {
         },
       }));
 
-      return api.callApi({ endpoint: 'resource' })
+      return callApi({ endpoint: 'resource' })
         .then((responseData) => {
           mockWindow.verify();
           expect(responseData).toEqual({});
@@ -122,7 +116,7 @@ describe(__filename, () => {
       const errorHandler = createStubErrorHandler();
       sinon.stub(errorHandler, 'handle');
 
-      return api.callApi({ endpoint: 'resource', errorHandler })
+      return callApi({ endpoint: 'resource', errorHandler })
         .then(unexpectedSuccess, () => {
           expect(errorHandler.handle.called).toBeTruthy();
           const args = errorHandler.handle.firstCall.args;
@@ -141,165 +135,51 @@ describe(__filename, () => {
         })
         .once()
         .returns(response);
-      return api.callApi({ endpoint: 'resource', method: 'GET' })
+      return callApi({ endpoint: 'resource', method: 'GET' })
         .then(() => mockWindow.verify());
     });
   });
 
   describe('makeQueryString', () => {
     it('transforms an object to a query string', () => {
-      const query = api.makeQueryString({ user: 123, addon: 321 });
+      const query = makeQueryString({ user: 123, addon: 321 });
       expect(query).toContain('user=123');
       expect(query).toContain('addon=321');
     });
 
     it('ignores undefined query string values', () => {
-      const query = api.makeQueryString({ user: undefined, addon: 321 });
+      const query = makeQueryString({ user: undefined, addon: 321 });
       expect(query).toEqual('?addon=321');
     });
 
     it('ignores null query string values', () => {
-      const query = api.makeQueryString({ user: null, addon: 321 });
+      const query = makeQueryString({ user: null, addon: 321 });
       expect(query).toEqual('?addon=321');
     });
 
     it('ignores empty string query string values', () => {
-      const query = api.makeQueryString({ user: '', addon: 321 });
+      const query = makeQueryString({ user: '', addon: 321 });
       expect(query).toEqual('?addon=321');
     });
 
     it('handles falsey integers', () => {
-      const query = api.makeQueryString({ some_flag: 0 });
+      const query = makeQueryString({ some_flag: 0 });
       expect(query).toEqual('?some_flag=0');
     });
 
     it('handles truthy integers', () => {
-      const query = api.makeQueryString({ some_flag: 1 });
+      const query = makeQueryString({ some_flag: 1 });
       expect(query).toEqual('?some_flag=1');
     });
 
     it('handles false values', () => {
-      const query = api.makeQueryString({ some_flag: false });
+      const query = makeQueryString({ some_flag: false });
       expect(query).toEqual('?some_flag=false');
     });
 
     it('handles true values', () => {
-      const query = api.makeQueryString({ some_flag: true });
+      const query = makeQueryString({ some_flag: true });
       expect(query).toEqual('?some_flag=true');
-    });
-  });
-
-  describe('featured add-ons api', () => {
-    const mockResponse = () => createApiResponse({
-      jsonData: {
-        results: [
-          { slug: 'foo' },
-          { slug: 'food' },
-          { slug: 'football' },
-        ],
-      },
-    });
-
-    it('sets the app, lang, and type query', () => {
-      mockWindow.expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/featured/?app=android&type=persona&lang=en-US`)
-        .once()
-        .returns(mockResponse());
-      return api.featured({
-        api: { clientApp: CLIENT_APP_ANDROID, lang: 'en-US' },
-        filters: { addonType: ADDON_TYPE_THEME },
-      })
-        .then((response) => {
-          expect(response).toEqual({
-            entities: {
-              addons: {
-                foo: { slug: 'foo' },
-                food: { slug: 'food' },
-                football: { slug: 'football' },
-              },
-            },
-            result: {
-              results: ['foo', 'food', 'football'],
-            },
-          });
-          return mockWindow.verify();
-        });
-    });
-  });
-
-  describe('add-on api', () => {
-    function mockResponse(responseProps = {}) {
-      return createApiResponse({
-        jsonData: {
-          name: 'Foo!',
-          slug: 'foo',
-        },
-        ...responseProps,
-      });
-    }
-
-    it('sets the lang and slug', () => {
-      mockWindow.expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/addon/foo/?lang=en-US`, {
-          body: undefined,
-          credentials: undefined,
-          headers: {},
-          method: 'GET',
-        })
-        .once()
-        .returns(mockResponse());
-      return api.fetchAddon({ api: { lang: 'en-US' }, slug: 'foo' })
-        .then(() => mockWindow.verify());
-    });
-
-    it('normalizes the response', () => {
-      mockWindow.expects('fetch').once().returns(mockResponse());
-      return api.fetchAddon('foo')
-        .then((results) => {
-          const foo = { slug: 'foo', name: 'Foo!' };
-          expect(results.result).toEqual('foo');
-          expect(results.entities).toEqual({ addons: { foo } });
-        });
-    });
-
-    it('fails when the add-on is not found', () => {
-      mockWindow
-        .expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/addon/foo/?lang=en-US`, {
-          body: undefined,
-          credentials: undefined,
-          headers: {},
-          method: 'GET',
-        })
-        .once()
-        .returns(mockResponse({ ok: false }));
-      return api.fetchAddon({ api: { lang: 'en-US' }, slug: 'foo' })
-        .then(unexpectedSuccess,
-          (error) => {
-            expect(error.message)
-              .toMatch(new RegExp('Error calling: /api/v3/addons/addon/foo/'));
-          });
-    });
-
-    it('includes the authorization token if available', () => {
-      const token = userAuthToken();
-      mockWindow
-        .expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/addon/bar/?lang=en-US`, {
-          body: undefined,
-          credentials: undefined,
-          headers: { authorization: `Bearer ${token}` },
-          method: 'GET',
-        })
-        .once()
-        .returns(mockResponse());
-      return api.fetchAddon({ api: { lang: 'en-US', token }, slug: 'bar' })
-        .then((results) => {
-          const foo = { slug: 'foo', name: 'Foo!' };
-          expect(results.result).toEqual('foo');
-          expect(results.entities).toEqual({ addons: { foo } });
-          mockWindow.verify();
-        });
     });
   });
 
@@ -307,7 +187,7 @@ describe(__filename, () => {
     function _createApiError({
       response = { status: 500 }, ...params } = {}
     ) {
-      return api.createApiError({ response, ...params });
+      return createApiError({ response, ...params });
     }
 
     it('includes an abbreviated URL', () => {
@@ -339,143 +219,6 @@ describe(__filename, () => {
       });
       expect(error.message)
         .toEqual('Error calling: /api/resource/ (status: 422)');
-    });
-  });
-
-  describe('login', () => {
-    const response = { token: userAuthToken() };
-    const mockResponse = () => createApiResponse({
-      jsonData: response,
-    });
-
-    it('sends the code and state', () => {
-      mockWindow
-        .expects('fetch')
-        .withArgs(`${apiHost}/api/v3/accounts/login/?lang=en-US`, {
-          body: '{"code":"my-code","state":"my-state"}',
-          credentials: 'include',
-          headers: { 'Content-type': 'application/json' },
-          method: 'POST',
-        })
-        .once()
-        .returns(mockResponse());
-      return api.login({ api: { lang: 'en-US' }, code: 'my-code', state: 'my-state' })
-        .then((apiResponse) => {
-          expect(apiResponse).toBe(response);
-          mockWindow.verify();
-        });
-    });
-
-    it('sends the config when set', () => {
-      sinon.stub(config, 'get').withArgs('fxaConfig').returns('my-config');
-      mockWindow
-        .expects('fetch')
-        .withArgs(`${apiHost}/api/v3/accounts/login/?config=my-config&lang=fr`)
-        .once()
-        .returns(mockResponse());
-      return api.login({ api: { lang: 'fr' }, code: 'my-code', state: 'my-state' })
-        .then(() => mockWindow.verify());
-    });
-  });
-
-  describe('startLoginUrl', () => {
-    const getStartLoginQs = (location) =>
-      querystring.parse(api.startLoginUrl({ location }).split('?')[1]);
-
-    it('includes the next path', () => {
-      const location = { pathname: '/foo', query: { bar: 'BAR' } };
-      expect(getStartLoginQs(location)).toEqual({ to: '/foo?bar=BAR' });
-    });
-
-    it('includes the next path the config if set', () => {
-      sinon.stub(config, 'get').withArgs('fxaConfig').returns('my-config');
-      const location = { pathname: '/foo' };
-      expect(getStartLoginQs(location)).toEqual({ to: '/foo', config: 'my-config' });
-    });
-  });
-
-  describe('categories api', () => {
-    function mockResponse(responseProps = {}) {
-      return createApiResponse({
-        jsonData: {
-          results: [
-            { slug: 'foo' },
-            { slug: 'food' },
-            { slug: 'football' },
-          ],
-        },
-        ...responseProps,
-      });
-    }
-
-    it('sets the lang and calls the right API endpoint', () => {
-      mockWindow.expects('fetch')
-        .withArgs(
-          `${apiHost}/api/v3/addons/categories/?lang=en-US`)
-        .once()
-        .returns(mockResponse());
-      return api.categories({
-        api: { clientApp: CLIENT_APP_ANDROID, lang: 'en-US' },
-      })
-        .then(() => mockWindow.verify());
-    });
-  });
-
-  describe('logOutFromServer', () => {
-    it('makes a delete request to the session endpoint', () => {
-      const mockResponse = createApiResponse({ jsonData: { ok: true } });
-      mockWindow.expects('fetch')
-        .withArgs(`${apiHost}/api/v3/accounts/session/?lang=en-US`, {
-          body: undefined,
-          credentials: 'include',
-          headers: { authorization: 'Bearer secret-token' },
-          method: 'DELETE',
-        })
-        .once()
-        .returns(mockResponse);
-      return api.logOutFromServer({ api: signedInApiState })
-        .then(() => mockWindow.verify());
-    });
-  });
-
-  describe('autocomplete api', () => {
-    const mockResponse = () => createApiResponse({
-      jsonData: {
-        results: [
-          createFakeAutocompleteResult({ name: 'foo' }),
-          createFakeAutocompleteResult({ name: 'food' }),
-          createFakeAutocompleteResult({ name: 'football' }),
-        ],
-      },
-    });
-
-    it('sets the app, lang, and query', () => {
-      mockWindow.expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/autocomplete/?app=android&q=foo&lang=en-US`)
-        .once()
-        .returns(mockResponse());
-      return api.autocomplete({
-        api: { clientApp: CLIENT_APP_ANDROID, lang: 'en-US' },
-        filters: {
-          query: 'foo',
-        },
-      })
-        .then(() => mockWindow.verify());
-    });
-
-    it('optionally takes addon type as filter', () => {
-      mockWindow.expects('fetch')
-        .withArgs(`${apiHost}/api/v3/addons/autocomplete/?app=android&q=foo&type=persona&lang=en-US`)
-        .once()
-        .returns(mockResponse());
-      return api.autocomplete({
-        api: { clientApp: CLIENT_APP_ANDROID, lang: 'en-US' },
-        filters: {
-          query: 'foo',
-          addonType: ADDON_TYPE_THEME,
-        },
-      })
-        .then(() => mockWindow.verify());
     });
   });
 });
